feat(domain): implement from UML
Adjust the controller to conform to the new format.
TODO: move the database logic into the repository

Adjust the trace importer v9 to conform to the new format and create
NetworkHosts.

Created TraceImportService as a central point for uniform
importing of traces.

Use scopes in NetFlow9TraceImporter.cs.

Delete TapasDatabase.cs as it wasn't being referenced anymore.

TAPAS-38