feat(fennec): save transport protocol to database
Now we read the Protocol (in Netflow9) and ProtocolIdentifier (in IPFIX) fields and write them to the database. These represent the transport layer protocol used in a session. Changed the TraceProtocol enum to support also an Unknown type in case we receive anything else than a UDP (17) or a TCP (6).

NETANOL-179