feat(fennec): add netflow5 support
* feat(fennec): netflow5 parser

Added NetFlowV5 support (tested). Testing was done via an external program which generates NetflowV5 export data. We do not have a NetflowV5 .pcap file with which we could test.

* minor(fennec): adapt netflow5 new multiplexer service

When creating NETANOL-198 branch n:m multiplexing still was not on dev. Had to add Netflow5 to be recognized as a protocol version.

* minor(fennec): log netflow v5

Logging for netflow v5 parser.

* fix(fennec): invalid dns add to cache

When we can't resolve an IP to a DNS name we still add it to the cache. This makes us not waste time with DNS resolving when we are not able to resolve a name.

* minor(fennec): fix log on invalid dns resolve

* minor(fennec): dns cache null when not resolved

If we fail to resolve a DNS we write null to DB.

* minor(fennec): change log level

An sflow capture file is included as well.

NETANOL-198